requirements.txt
openai>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
anthropic>=0.40.0
httpx>=0.25.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0
pyyaml>=6.0.1
python-multipart>=0.0.6
google-api-python-client>=2.100.0
google-auth>=2.23.0
jinja2>=3.1.2
itsdangerous>=2.1.2
